requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
WTForms==3.1.1
Werkzeug==3.0.1
anthropic==0.18.1
python-dotenv==1.0.0
cryptography==41.0.7
python-socketio==5.11.0
email-validator==2.1.0
